pyproject.toml
[project]
name = "asm-notebook"
version = "0.1.0"
description = ""
authors = [
    {name = "Charley Thomas",email = "charley.thomas1@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "typer (>=0.24.0,<0.25.0)",
    "rich (>=14.3.3,<15.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "dnspython (>=2.8.0,<3.0.0)",
    "diskcache (>=5.6.3,<6.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "fastapi (>=0.129.2,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "ipwhois (>=1.3.0,<2.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "google-auth (>=2.40.0,<3.0.0)",
    "requests (>=2.32.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.2,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=24.10.0,<25.0.0)"
]


[tool.poetry.scripts]
asmnb = "asm_notebook.cli:main"
